Move out retlist construction.
